requirements.txt
requests
bs4
pandas
openpyxl
# react-form-wizard-component
sqlalchemy
python-dotenv
pymysql
# react-date-range
# react-hook-form
flask
flask_cors
# styled-components
# @tanstack/react-table